cargo-client-api add pagination

+++ cargo-client-api/src/main/java/com/example/cargoClientApi/repository/SensorEventRepository.java
@@ -3,5 +3,5 @@ package com.example.cargoClientApi.repository;
 import com.example.cargoClientApi.models.SensorEvent;
 import org.springframework.data.cassandra.repository.CassandraRepository;
 
-public interface SensorEventRepository extends CassandraRepository<SensorEvent, Integer> {
+public interface SensorEventRepository extends CassandraRepository<SensorEvent, Long> {
 }
